requirements.txt
sanic
pydantic
sqlalchemy
psycopg2-binary
gunicorn
python-dotenv
SQLAlchemy
asyncpg
